feat: реализованы функционал Booking(Controller, Service, Repository).

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -3,15 +3,16 @@ package ru.practicum.shareit.user.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ru.practicum.shareit.exceptions.ConflictException;
+import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.shareit.exceptions.ObjectNotFoundException;
-import ru.practicum.shareit.exceptions.ValidationException;
-import ru.practicum.shareit.user.UserRepository;
+import ru.practicum.shareit.user.UserMapper;
+import ru.practicum.shareit.user.dao.UserRepository;
 import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.model.User;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
 import static ru.practicum.shareit.user.UserMapper.toUser;
 import static ru.practicum.shareit.user.UserMapper.toUserDto;
@@ -24,45 +25,33 @@ public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
 
     @Override
+    @Transactional
     public UserDto addNewUser(UserDto userDto) {
-        userEmailDuplicateCheck(userDto);
         User user = toUser(userDto);
-
         return toUserDto(userRepository.save(user));
     }
 
     @Override
+    @Transactional
     public UserDto updateUser(UserDto userDto, Long userId) {
-        userRepository.findById(userId).orElseThrow(
+        User userFromStorage = userRepository.findById(userId).orElseThrow(
                 () -> {
                     log.debug("User with id {} was not found.", userId);
                     throw new ObjectNotFoundException(String.format("User with id: %s was not found!", userId));
                 }
         );
 
-        User user = toUser(userDto);
-        UserDto userFromStorage = getUserById(userId);
-
-        if (Objects.isNull(userDto.getName())) {
-            user.setName(userFromStorage.getName());
+        if (!Objects.isNull(userDto.getEmail())) {
+            userFromStorage.setEmail(userDto.getEmail());
         }
-
-        if (Objects.isNull(userDto.getEmail())) {
-            user.setEmail(userFromStorage.getEmail());
-        } else {
-            String email = userFromStorage.getEmail();
-            boolean isEmailNotChanged = email.equals(userDto.getEmail());
-
-            if (!isEmailNotChanged) {
-                userEmailDuplicateCheck(userDto);
-            }
+        if (!Objects.isNull(userDto.getName())) {
+            userFromStorage.setName(userDto.getName());
         }
-        user.setId(userId);
-        userValidation(user);
-        return toUserDto(userRepository.save(user));
+        return toUserDto(userRepository.save(userFromStorage));
     }
 
     @Override
+    @Transactional(readOnly = true)
     public UserDto getUserById(Long userId) {
         User user = userRepository.findById(userId).orElseThrow(
                 () -> {
@@ -74,38 +63,19 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<UserDto> getAllUsers() {
-        return toUserDto(userRepository.findAll());
+        return userRepository.findAll().stream().map(UserMapper::toUserDto).collect(Collectors.toList());
     }
 
     @Override
+    @Transactional
     public void deleteUser(Long userId) {
-        userRepository.deleteById(userId);
-    }
-
-    private void userEmailDuplicateCheck(UserDto user) {
-        if (userRepository.getAllEmails().contains(user.getEmail())) {
-            log.debug("User email {} already exist.", user.getEmail());
-            throw new ConflictException(String.format("User email %s already exist.", user.getEmail()));
-        }
-    }
-
-    private void userValidation(User user) {
-        if (user.getEmail().isBlank()) {
-            log.debug("User with email {} was attempted to create.", user.getEmail());
-            throw new ValidationException("User email cannot be empty.");
-        }
-        if (!user.getEmail().contains("@")) {
-            log.debug("User with email {} was attempted to create.", user.getEmail());
-            throw new ValidationException("User email have to contain '@' sign.");
-        }
-        if (user.getName().isBlank()) {
-            log.debug("User with login {} was attempted to create.", user.getName());
-            throw new ValidationException("User login cannot be empty or blank.");
-        }
-        if (user.getName().contains(" ")) {
-            log.debug("User with login {} was attempted to create.", user.getName());
-            throw new ValidationException("User login cannot have spaces in it.");
+        if (userRepository.findById(userId).isPresent()) {
+            userRepository.deleteById(userId);
+        } else {
+            log.debug("User with id {} was not found.", userId);
+            throw new ObjectNotFoundException(String.format("User with id: %s was not found!", userId));
         }
     }
 }
